Add 0% item detection
