+++ src/main/java/com/wynntils/modules/utilities/configs/UtilitiesConfig.java
@@ -131,7 +131,7 @@ public class UtilitiesConfig extends SettingsClass {
         @Setting(displayName = "Legacy Identification Values", description = "Should mana and life steal stats display as /4s instead of /3s and /5s?")
         public boolean legacyIds = false;
 
-        @Setting(displayName = "Rainbow Perfect Items", description = "Should perfect items have rainbow names?")
+        @Setting(displayName = "Animate Perfect/Defect Item Names", description = "Should perfect and defective items have animated names?\n\n§8 This includes perfect items having rainbow names.")
         public boolean rainbowPerfect = true;
 
         @Setting(displayName = "Categorize Identifications", description = "Should the identifications in an item's tooltip be categorized?")
